Server cleanup and prototype client
Cleaned up the server behavior for alarm command and telemetry. Also,
implemented a shutdown hook to make it behave more like a linux
service. Included alarm-server.sh script so that it can launch at boot
and behave like a linux daemon. And included an LCARS iPhone keypad
prototype interface, cacheable, and set up to run as a home screen app.